requirements.txt
horovod==0.28.1
matplotlib==3.8.0
numpy==1.22.4
open_clip_torch==2.16.0
Pillow==10.3.0
scikit_learn==1.3.0
scipy==1.13.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
requests
